move `AnalyticalTableColumnDefinition` to index, add `UI5DomRef` types
