+++ docs/2-MigrationGuide.stories.mdx
@@ -21,7 +21,7 @@ or the [changelog](https://github.com/SAP/ui5-webcomponents-react/blob/main/CHAN
 
 ### Changed TypeScript `interfaces` for `ref`
 
-The TypeScript `interfaces` that can be used for accessing the DOM-`ref` of a component have been updated:
+TypeScript `interfaces` that can be used for accessing the DOM-`ref` of a component have been updated:
 
 - The `Ui5` prefix has been removed
 - The `interfaces` can now be imported from the component itself
